requirements.txt
langgraph==0.6.6
langchain==0.3.27
pydantic==2.9.2
python-dotenv==1.0.1
ollama==0.3.2

typer==0.12.5
rich==13.8.1
click==8.1.7
black==24.8.0
isort==5.13.2
pytest==8.3.3
